server refactoring
