+++ server/models/category.model.js
@@ -1,0 +1,13 @@
+const categorySchema = {
+  categoryId: "",
+  parentCategoryId: "",
+  categoryTitle: "",
+  _isDeleted: false,
+  _createdAt: 0,
+  _deletedAt: 0,
+  lvl: 0,
+}
+
+module.exports = {
+  categorySchema,
+};
