+++ server/models/task.model.js
@@ -1,0 +1,12 @@
+const taskSchema = {
+  categoryId: "",
+  taskId: "",
+  title: "",
+  description: "",
+  _createdAt: 0,
+  status: "new",
+};
+
+module.exports = {
+  taskSchema
+};
