+++ server/routes/tasks.routes.js
@@ -1,0 +1,13 @@
+const express = require("express");
+const router = express.Router();
+
+const tasksController = require("../controllers/tasks.controller");
+
+router.get("/", tasksController.get);
+
+router.post("/", tasksController.post);
+
+router.patch("/", tasksController.patch);
+
+module.exports = router;
+
