+++ src/components/EditTask/EditTask.scss
@@ -1,6 +1,6 @@
 .EditTask {
-display: flex;
-flex-direction: column;
+  display: flex;
+  flex-direction: column;
 }
 
 .EditTask-buttons {
